83bb063 Merge topic 'wrap-python-python-h' into master
09a7973 COMP: Do not include vtkPython.h in vtkWrapPython.c.
11c2748 Merge topic 'fix-msvc-compiler-warnings' into master
a6de1b3 Fix msvc compiler warning in particle tracers
0d55b06 Merge topic 'cleanup-compilation-warnings' into master
a824648 Fix dashboard compilation warnings.
6e9aa0f Merge topic 'MathTextPrintFix' into master
a7d1a05 Ensure that MathTextActor's members are fully initialized.
7c50585 Merge topic 'add_execution_timer' into master
fd2fe65 Merge topic 'wrapper-parser-enhancements' into master
4effb0f Merge topic 'vtkXdmfReader-TestSetGet' into master
d6bab51 vtkXdmfReader: Do not crash in GetSIL without active domain
f7f494f Merge topic 'AddTests-Views/Infovis/Testing/Cxx' into master
e83f010 Merge topic 'module-tcl-names' into master
a93b260 vtkExecutionTimer: Add missing comment to satisfy HeaderTest
9ff30a7 Use TCL_NAME for VTK Tcl static package list
...

+++ Filters/ParallelGeometry/vtkPStructuredGridConnectivity.cxx
@@ -921,7 +921,7 @@ void vtkPStructuredGridConnectivity::DeserializeDataArray(
       float *data       = NULL;
       bytestream.Pop( data, size );
       assert( "pre: de-serialized data array is not of the expected size" &&
-              (size == (numberOfTuples*numberOfComponents) ) );
+              (size == static_cast<unsigned int>((numberOfTuples*numberOfComponents)) ) );
 
       // STEP 1: Allocate vtkdata array
       dataArray = vtkDataArray::CreateDataArray( dataType );
@@ -948,7 +948,7 @@ void vtkPStructuredGridConnectivity::DeserializeDataArray(
       double *data      = NULL;
       bytestream.Pop(data,size);
       assert( "pre: de-serialized data array is not of the expected size" &&
-              (size==(numberOfTuples*numberOfComponents) ) );
+              (size == static_cast<unsigned int>((numberOfTuples*numberOfComponents)) ) );
 
       // STEP 1: Allocate vtkdata array
       dataArray = vtkDataArray::CreateDataArray( dataType );
@@ -975,7 +975,7 @@ void vtkPStructuredGridConnectivity::DeserializeDataArray(
       int *data         = NULL;
       bytestream.Pop(data,size);
       assert( "pre: de-serialized data array is not of the expected size" &&
-              (size==(numberOfTuples*numberOfComponents) ) );
+              (size == static_cast<unsigned int>((numberOfTuples*numberOfComponents)) ) );
 
       // STEP 1: Allocate vtkdata array
       dataArray = vtkDataArray::CreateDataArray( dataType );
